requirements.txt
SQLAlchemy==1.1.10
APScheduler==3.3.1
paramiko==2.1.2
peewee==2.6.3
mysqlclient==1.3.12
requests==2.18.4
scp==0.10.2
django-extensions==2.0.6
celery==4.1.0
redis==2.10.6